requirements.txt
fastapi==0.92.0
SQLAlchemy==2.0.4
python-dotenv==1.0.0
alembic==1.9.4
#psycopg2-binary==2.9.5
uvicorn==0.20.0
pydantic==1.10.5
databases[postgresql]==0.3.1
#poetry==1.4.0
